labs/06-observability/demo-04-langchain-auto-tracing/pyproject.toml
[project]
name = "demo-04-langchain-auto-tracing"
version = "0.1.0"
description = "Automatic LangSmith tracing for LangChain chains, agents, and LCEL pipelines"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langsmith>=0.1.0",
    "langchain>=1.2.0",
    "langchain-openai>=0.2.0",
    "langchain-core>=0.3.0",
    "python-dotenv>=1.0.0",
]
